Add skeleton implementation for the External logging extension point
