+++ pom.xml
@@ -9,7 +9,7 @@
   </parent>
   
   <properties>
-    <jenkins.version>1.642.3</jenkins.version>
+    <jenkins.version>2.23-SNAPSHOT</jenkins.version>
     <java.level>7</java.level>
   </properties>
